+++ src/server/builder.rs
@@ -16,7 +16,7 @@ use super::KeepAlive;
 pub(crate) trait ServiceProvider {
     fn register(
         &self, server: Server, lst: net::TcpListener, host: String,
-        addr: net::SocketAddr, keep_alive: KeepAlive, client_timeout: u64,
+        addr: net::SocketAddr, keep_alive: KeepAlive, secure: bool, client_timeout: u64,
         client_shutdown: u64,
     ) -> Server;
 }
@@ -28,7 +28,6 @@ where
 {
     factory: F,
     acceptor: A,
-    no_client_timer: bool,
 }
 
 impl<F, H, A> HttpServiceBuilder<F, H, A>
@@ -40,27 +39,13 @@ where
 {
     /// Create http service builder
     pub fn new(factory: F, acceptor: A) -> Self {
-        Self {
-            factory,
-            acceptor,
-            no_client_timer: false,
-        }
-    }
-
-    pub(crate) fn no_client_timer(mut self) -> Self {
-        self.no_client_timer = true;
-        self
+        Self { factory, acceptor }
     }
 
     fn finish(
-        &self, host: String, addr: net::SocketAddr, keep_alive: KeepAlive,
+        &self, host: String, addr: net::SocketAddr, keep_alive: KeepAlive, secure: bool,
         client_timeout: u64, client_shutdown: u64,
     ) -> impl ServiceFactory {
-        let timeout = if self.no_client_timer {
-            0
-        } else {
-            client_timeout
-        };
         let factory = self.factory.clone();
         let acceptor = self.acceptor.clone();
         move || {
@@ -68,12 +53,12 @@ where
             let settings = WorkerSettings::new(
                 app,
                 keep_alive,
-                timeout as u64,
+                client_timeout,
                 client_shutdown,
                 ServerSettings::new(addr, &host, false),
             );
 
-            if timeout == 0 {
+            if secure {
                 Either::A(ServerMessageAcceptor::new(
                     settings.clone(),
                     TcpAcceptor::new(acceptor.create().map_err(AcceptorError::Service))
@@ -88,14 +73,16 @@ where
             } else {
                 Either::B(ServerMessageAcceptor::new(
                     settings.clone(),
-                    TcpAcceptor::new(AcceptorTimeout::new(timeout, acceptor.create()))
-                        .map_err(|_| ())
-                        .map_init_err(|_| ())
-                        .and_then(
-                            HttpService::new(settings)
-                                .map_init_err(|_| ())
-                                .map_err(|_| ()),
-                        ),
+                    TcpAcceptor::new(AcceptorTimeout::new(
+                        client_timeout,
+                        acceptor.create(),
+                    )).map_err(|_| ())
+                    .map_init_err(|_| ())
+                    .and_then(
+                        HttpService::new(settings)
+                            .map_init_err(|_| ())
+                            .map_err(|_| ()),
+                    ),
                 ))
             }
         }
@@ -112,7 +99,6 @@ where
         HttpServiceBuilder {
             factory: self.factory.clone(),
             acceptor: self.acceptor.clone(),
-            no_client_timer: self.no_client_timer,
         }
     }
 }
@@ -126,13 +112,20 @@ where
 {
     fn register(
         &self, server: Server, lst: net::TcpListener, host: String,
-        addr: net::SocketAddr, keep_alive: KeepAlive, client_timeout: u64,
+        addr: net::SocketAddr, keep_alive: KeepAlive, secure: bool, client_timeout: u64,
         client_shutdown: u64,
     ) -> Server {
         server.listen2(
             "actix-web",
             lst,
-            self.finish(host, addr, keep_alive, client_timeout, client_shutdown),
+            self.finish(
+                host,
+                addr,
+                keep_alive,
+                secure,
+                client_timeout,
+                client_shutdown,
+            ),
         )
     }
 }